filter: Add '*/manual/*' and '*/autom4te.cache/*'
Add the terms to the find_filter.
Files in the source tree with these names will not arm the build system when
they are updated.
Reason: they are generated files, created when building a platform.
Thus, if you build platform 1, creating these files will arm a build of
platform 2, even when platform 2 is up to date.
You would certainly die before the build completes.
This problem is known to occur in glibc.

Keep an eye out to see if the file 'configure' has the same problem.
I am not protecting against that now, just to make sure that I don't introduce
a change having unintended or unexpected consequences.