Fix xx-configure and filter all-cfg.texi.
Previously a change to "configure" in the source tree was not picked up until
the build phase (of configure, build and install phases.)
Now it will kick off the configure script during the configure phase,
whenever possible.

Adding '*/doc/all-cfg.texi' to the find filter because gdb building throws
a file with that name back into the source tree.