Flexible prefix/DESTDIR preparation.
A set of variables is declared to allow for a variety of prefix/DESTDIR
usages.

Typically there is one prefix, and DESTDIR may or may not be used.
DESTDIR is common when building embedded systems (cross compilation.)
Ebuild has not used DESTDIR. It has used a separate prefix per package.
For example, the bash prefix is /scr/bsi/qya/ope/bui/install-ppc7450/bash;
the ncurses prefix is /scr/bsi/qya/ope/bui/install-ppc7450/ncurses.
I have a variable PER_PACKAGE_PREFIX; you set it for the above behavior.
I also provide a PER_PACKAGE_DESTDIR control variable.
It is unlikely that both would be set; and in a standard autotools system,
neither would be set.

I provide a set of variables for each platform.
Imagine someone wants to alter the default behavior so that she can work on
qnative, and assume it is likely they only want to alter one platform;
what I have is most convenient for that mode of developing.

Normally one would adjust the PREFIX_BASE for his platform and have LIBDIR
and friends follow along.
We are set up to make that easy.
Note that in a standard system PREFIX_BASE is equal to prefix
(as in configure --prefix); they differ only if PER_PACKAGE_PREFIX is set.
Likewise with DESTDIR_BASE.

My first commit has the variables set to the traditional ebuild style (as
opposed to the traditional autotools style.)
This is just to avoid a bumpy transition.
One peculiarity of the ebuild style is that prefix is set to where I am
"staging" a build and these strange prefixes end up in the programs that are
distributed on the router.
It has no meaning for performance, but can be confusing.
But the code can be executed from the prefix location on the router which is
immensely useful for NFS mounted "prefix" directories.
The same behavior is obtained when using a real prefix --
but in that case we NFS mount the build tree.
That again is standard autotools methodology.

The use of multiple prefixes and/or DESTDIRs can create difficulties for
developers of complex open source code that has been developed with traditional
autotools techniques.
For this and other reasons, I recommend the singles... after "the bump."

This infrastructure is being committed prior to the changes that activate it.
Some variable seen here come from the follow on commit.
Staging the changes to make things simpler.
I have attempted to use upper case variable for those that you might change;
lower case for those that are "calculated."