pkgPhaseDependMacro introduction.
This is nothing more than a macro, for things like the common expression
pkg1_configure_depend = pkg2-install ... pkgn-install.

The calling sequence is:
pkg1_depend = pkg2 ... pkgn
$(call pkgPhaseDependMacro,pkg1)

The macro committed here expands to the original style dependencies (to avoid a
bump), but the intention is for it to expand to:
pkg1_configure_depend = pkg2-configure ... pkgn-configure
pkg1_build_depend = pkg2-build ... pkgn-build
pkg1_install_depend = pkg2-install... pkgn-install

whence it will not be universally usable, but usable in most cases.
(With enough makefile work in individual packages it is universally usable.)

I am committing it now to allow use of the macro by the .mk files;
many of which will be committed shortly.