New variables for prefix and DESTDIR...
They are used currently in a few .mk files but are not used in the ebuild
control structure yet.

Later I plan to deprecate INSTALL_DIR and simply replace it with
DFLT_INSTALL_DIR for clarity.

The planned follow on changes for this Makefile will make it possible to use
the adjustable prefix and DESTDIR that all this work has been leading up to.
The other follow on will use the new version of pkgPhaseDependMacro that allows
one to build with very little install phase activity.
And there will be even less install phase activity if work is done on the
indivdual package Makefiles or code files to not require things from the
install tree.
That will speed up the build cycle for something like routing-sw or qvnet.